✨: Project Restructuring
