+++ tic-rs-client/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tic-rs-client"
+authors = ["Awiteb github.com/TheAwiteb"]
+description = "tic-rs client"
+edition = "2021"
+repository = "https://github.com/TheAwiteb/tic-rs"
+rust-version = "1.60" # 7737e0b5c 2022-04-04
+version = "0.1.0"
+
+[dependencies]
+clap = { version = "3.1.18", features = ["derive"] }
+promptly = "0.3.0"
+colored = "2.0.0"
+tictactoy = { path = "../tictactoy" }
