+++ tic-rs-client/src/utils.rs
@@ -1,7 +1,6 @@
-use super::cli::CliApp;
-///                       Rust Tic Tac Toy (x/o)
-///                 Copyright (C) 2020-2022  TheAwiteb
-///                 https://github.com/TheAwiteb/tic-rs
+/// Rust Tic Tac Toy (x/o)
+/// Copyright (C) 2020-2022  TheAwiteb
+/// <https://github.com/TheAwiteb/tic-rs>
 ///
 /// This program is free software: you can redistribute it and/or modify it under
 /// the terms of the GNU Affero General Public License as published by the Free
@@ -15,9 +14,10 @@ use super::cli::CliApp;
 ///
 /// You should have received a copy of the GNU Affero General Public License along
 /// with this program.  If not, see <http://www.gnu.org/licenses/>.
-use super::tictactoy::player::{PlayTypes, Player, Pointers};
+use super::cli::CliApp;
 use colored::Colorize;
 use promptly::{prompt, ReadlineError};
+use tictactoy::player::{PlayTypes, Player, Pointers};
 
 /// Ask user to add player
 pub fn get_player(pointer: Pointers, app: &CliApp) -> Result<Player, ReadlineError> {
